Store D-ID and OPEN AI Api Keys in cookies
Changed the app to store D-ID and OPEN AI Api Keys in cookies.  This allows the app to be hosted as a demo application where individual users can obtain their own API keys to test with out having to share a common pool of requests.  It prevents rampant abuse of potentially skyrocketing costs and allows each user to track charges or credits.